code refactor

+++ src/main/java/Simulation/Output/IInput.java
@@ -3,8 +3,8 @@ package Simulation.Output;
 import java.util.List;
 
 public interface IInput {
-    int loadPopulation();
-    int loadWealth();
-    int loadCombatCapability();
+    long loadPopulation();
+    long loadWealth();
+    long loadCombatCapability();
     List<String> getRandomEventNames();
 }
